add: type controller

+++ src/main/java/pokemon/server/repository/TypeRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import pokemon.server.model.Type;
 
-public interface TypeRepository extends JpaRepository<Type, Long> {
+public interface TypeRepository extends JpaRepository<Type, Integer> {
 
 }
